deps(zod): move to types
