+++ apps/bot/src/commands/prefix.ts
@@ -1,4 +1,4 @@
-import { db, Prefix } from "database";
+import { db, Prefix as DBPrefix } from "database";
 import { ApplicationCommandOptionType, ApplicationCommandType, ChatInputCommandInteraction, CommandInteraction, Message, userMention } from "discord.js";
 import { cache } from "..";
 import { CommandSource } from "../../types/Command";
@@ -8,7 +8,7 @@ import { sliceCommand } from "../utils/commands/slice";
 const commandName = "prefix" as const;
 const valueOption = "value";
 
-const storePrefix = ({ guildId, prefix, userId }: Prefix) =>
+const storePrefix = ({ guildId, prefix, userId }: DBPrefix) =>
     db.prefix.update({
         where: { guildId },
         include: { guild: true },
