+++ package.json
@@ -25,7 +25,6 @@
   "dependencies": {
     "custom-env": "^2.0.1",
     "dotenv": "^16.0.3",
-    "find-config": "^1.0.0",
-    "zod": "^3.21.4"
+    "find-config": "^1.0.0"
   }
 }
