+++ packages/types/Prefix.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+
+export const prefixObject = z.object({
+    guildId: z.string(),
+    userId: z.string(),
+    prefix: z.string()
+});
+
+export type Prefix = z.infer<typeof prefixObject>
